123-ajuda-api-main/123-ajuda-api-main/requirements.txt
attrs==23.1.0
blinker==1.6.2
click==8.1.6
flasgger==0.9.7b2
flask==2.3.2
docker-compose==1.29.2
flask-sqlalchemy==3.0.5
greenlet==2.0.2
gunicorn==21.2.0
importlib-metadata==6.8.0
importlib-resources==6.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema>=3.0.1,<4.0
jsonschema-specifications==2023.7.1
MarkupSafe==2.1.3
mistune==3.0.1
mysql-connector-python==8.1.0
packaging==23.1
pkgutil-resolve-name==1.3.10
protobuf==4.21.12
python-dotenv>=0.13.0,<1.0.0
PyYAML>=3.10,<6.0
referencing==0.30.2
rpds-py==0.9.2
six==1.16.0
SQLAlchemy==2.0.19
typing-extensions==4.7.1
Werkzeug==2.3.6
zipp==3.16.2
